SQLHandler instance update
SQLHandler now holds a static instance to make the SQLHandler available for the whole program. It must be this way because SQLHandler must be instanciable so that FormSettings can use a temporary SQLHandler to get the available databases without overwriting the globaly used SQLHandler instance.

GoalManager now uses SQLHandler.Current instead of own local instance.
Program now generates the SQLHandler instance for SQLHandler.Current on start.